requirements.txt
requests
kafka-python
psycopg2-binary
pandas
scikit-learn
joblib
